requirements.txt

colorama==0.4.6
Flask==3.0.3
Flask-RESTful==0.3.10

future==1.0.0

google-api-core==2.24.2
google-api-python-client==2.151.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.69.1
grpcio==1.71.0
gspread==6.1.4
huggingface-hub==0.29.3
keras==3.9.0

matplotlib==3.9.4
nltk==3.9.1
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2

pandas==2.2.3
pillow==11.0.0
plotly==6.0.0
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
streamlit==1.43.2
tensorboard==2.19.0
tensorflow==2.19.0
tokenizers==0.21.1
torch==2.6.0
transformers==4.49.0
urllib3==2.3.0
watchdog==6.0.0
zipp==3.21.0
